backend/requirements.txt
supabase
python-dotenv
requests
beautifulsoup4
openai
pillow
feedparser
google-genai
